### 2018-02-01 #563 @whedon
[PRE REVIEW]: checkr: An R package for Assertive Programming
**Submitting author:** @joethorley (<a href="http://orcid.org/0000-0002-7683-4592">Joseph Thorley</a>)
**Repository:** <a href="https://github.com/poissonconsulting/checkr" target ="_blank">https://github.com/poissonconsulting/checkr</a>
**Version:** v0.0.2
