### 2018-02-06 2731678
- changed `check_join()` error message to     *...join in x and y must include all the rows in x* as opposed to     *...join in x and y violates referential integrity*
