+++ R/join.R
@@ -2,10 +2,13 @@
 #' 
 #' Checks that the columns in data frame x form a many-to-one
 #' join with the corresponding columns in y.
+#' That is to say the join is a unique key in y and all the rows in
+#' x have a match in y.
 #'
 #' @param x The object to check.
 #' @param y The parent data frame.
 #' @param by A character vector or named character vector of the columns to join by. 
+#' @param all A flag indicating whether all the rows in y should have a match in x.
 #' @param x_name A string of the name of the object x.
 #' @param y_name A string of the name of the object y.
 #' @param error A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.
@@ -19,6 +22,7 @@
 #' check_join(data1, data2, by = c(x = "y"), error = FALSE)
 check_join <- function(x, y,
                        by = NULL,
+                       all = FALSE,
                        x_name = substitute(x),
                        y_name = substitute(y),
                        error = TRUE) {
@@ -29,7 +33,8 @@ check_join <- function(x, y,
   check_data(y, x_name = y_name)
   check_flag_internal(error)
   
-  checkor(check_null(by), check_vector(by, "", length = c(1, Inf)))
+  checkor(check_null(by), check_vector(by, "", length = TRUE))
+  check_flag(all)
   
   if(is.null(by)) {
     by <- intersect(colnames(x), colnames(y))
@@ -45,20 +50,26 @@ check_join <- function(x, y,
 
   check_key(y, by, x_name = y_name, error = error)
   
-  if(!nrow(x)) return(invisible(x))
+  if(!all && !nrow(x)) return(invisible(x))
   
   check_missing_colnames(x, "..ID")
   check_missing_colnames(y, "..ID")
   
-  ..ID <- 1:nrow(x)
+  ..ID.x <- 1:nrow(x)
+  ..ID.y <- 1:nrow(y)
   
-  x$..ID <- ..ID
+  x$..ID.x <- ..ID.x
+  y$..ID.y <- ..ID.y
   
   suppressWarnings(z <- merge(x, y, by.x = names(by), by.y = by))
   
-  if(!identical(sort(unique(z$..ID)), sort(..ID))) {
-        on_fail("join between ", x_name, " and ", y_name, " violates referential integrity", error = error)
+  if(!identical(sort(unique(z$..ID.x)), sort(..ID.x))) {
+        on_fail("join between ", x_name, " and ", y_name, " must include all the rows in ", x_name, error = error)
   }
-  x$..ID <- NULL
+  if(all && !identical(sort(unique(z$..ID.y)), sort(..ID.y))) {
+        on_fail("join between ", x_name, " and ", y_name, " must include all the rows in ", y_name, error = error)
+  }
+  
+  x$..ID.x <- NULL
   invisible(x)
 }
