### 2018-02-06 79facd3
rename all = TRUE to all_y in check_join
