### 2018-02-28 f15ee2d
- check_vector() and check_list() now allow `named` argument to be a regular expression - added `regex = ".*"` argument to `check_named()` - added `check_regex()` function
