+++ README.md
@@ -159,7 +159,7 @@ y %>% check_list(x_name = "y")
 To cite package 'checkr' in publications use:
 
   Joe Thorley (2018). checkr: Check the Properties of Common R
-  Objects. R package version 0.0.2.9003.
+  Objects. R package version 0.0.2.9004.
   https://github.com/poissonconsulting/checkr
 
 A BibTeX entry for LaTeX users is
@@ -168,7 +168,7 @@ A BibTeX entry for LaTeX users is
     title = {checkr: Check the Properties of Common R Objects},
     author = {Joe Thorley},
     year = {2018},
-    note = {R package version 0.0.2.9003},
+    note = {R package version 0.0.2.9004},
     url = {https://github.com/poissonconsulting/checkr},
   }
 ```
