+++ man/check_regex.Rd
@@ -1,0 +1,27 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/regex.R
+\name{check_regex}
+\alias{check_regex}
+\title{Check Matches Regular Expression}
+\usage{
+check_regex(x, regex = ".*", x_name = substitute(x), error = TRUE)
+}
+\arguments{
+\item{x}{The object to check.}
+
+\item{regex}{A string of the regular expression.}
+
+\item{x_name}{A string of the name of the object.}
+
+\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
+}
+\value{
+An invisible copy of x (if it doesn't throw an error).
+}
+\description{
+Check Matches Regular Expression
+}
+\examples{
+check_regex("foo", "fo")
+check_regex("foo", "fo$", error = FALSE)
+}
