### 2018-02-28 8d3f59a
added `check_nchar()` function
