+++ man/check_nchar.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/nchar.R
+\name{check_nchar}
+\alias{check_nchar}
+\title{Check Number of Characters}
+\usage{
+check_nchar(x, nchar = TRUE, x_name = substitute(x), error = TRUE)
+}
+\arguments{
+\item{x}{The object to check.}
+
+\item{nchar}{A flag indicating whether x should have characters or a missing value indicating no requirements or a count or count range of the number of characters.}
+
+\item{x_name}{A string of the name of the object.}
+
+\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
+}
+\value{
+An invisible copy of x (if it doesn't throw an error).
+}
+\description{
+Check Number of Characters
+}
+\examples{
+check_nchar(c("foo", "bar"), nchar = 3)
+}
