+++ NAMESPACE
@@ -20,6 +20,7 @@ export(check_missing_colnames)
 export(check_missing_names)
 export(check_named)
 export(check_names)
+export(check_nchar)
 export(check_ncol)
 export(check_nlevels)
 export(check_nrow)
