### 2018-02-28 3100c0f
- `check_vector()` and `check_list()` now allow `named` argument to be a count range - added `nchar = c(0L, .Machine$max.integer)` argument to `check_named()`
#4
