### 2018-02-28 f1d855d
drop "elements of" from check_nchar() error message
