+++ tests/testthat/test-vector.R
@@ -10,6 +10,6 @@ test_that("vector", {
   expect_error(check_vector(vec, named = "x"), "names of vec must match regular expression 'x'")
   
   vec <- c(y = 1, y = 2)
-  expect_error(check_vector(vec, named = 2), "elements of names of vec must have 2 characters")
+  expect_error(check_vector(vec, named = 2), "names of vec must have 2 characters")
   expect_error(check_vector(vec, named = FALSE), "vec must be unnamed")
 })
