+++ tests/testthat/test-named.R
@@ -11,7 +11,7 @@ test_that("named", {
   expect_error(check_named(vec, regex = "x"), "names of vec must match regular expression 'x'")
   
   vec <- c(y = 1, yy = 2)
-  expect_error(check_named(vec, nchar = 1), "elements of names of vec must have 1 character")
-  expect_error(check_named(vec, nchar = c(2,4)), "elements of names of vec must have at least 2 characters")
+  expect_error(check_named(vec, nchar = 1), "names of vec must have 1 character")
+  expect_error(check_named(vec, nchar = c(2,4)), "names of vec must have at least 2 characters")
   expect_identical(check_named(vec, nchar = c(1,2)), vec)
 })
