+++ README.md
@@ -152,6 +152,26 @@ y %>% check_list(x_name = "y")
 #> Error: y must be a list
 ```
 
+## Installation
+
+To install the latest official release from
+[CRAN](https://CRAN.R-project.org/package=checkr)
+
+    install.packages("checkr")
+
+To install the latest development version from
+[GitHub](https://github.com/poissonconsulting/checkr)
+
+    # install.packages("devtools")
+    devtools::install_github("poissonconsulting/checkr")
+
+To install the latest development version from the Poisson drat
+[repository](https://github.com/poissonconsulting/drat)
+
+    # install.packages("drat")
+    drat::addRepo("poissonconsulting")
+    install.packages("checkr")
+
 ## Citation
 
 ``` 
@@ -159,7 +179,7 @@ y %>% check_list(x_name = "y")
 To cite package 'checkr' in publications use:
 
   Joe Thorley (2018). checkr: Check the Properties of Common R
-  Objects. R package version 0.0.2.9006.
+  Objects. R package version 0.1.0.
   https://github.com/poissonconsulting/checkr
 
 A BibTeX entry for LaTeX users is
@@ -168,7 +188,7 @@ A BibTeX entry for LaTeX users is
     title = {checkr: Check the Properties of Common R Objects},
     author = {Joe Thorley},
     year = {2018},
-    note = {R package version 0.0.2.9006},
+    note = {R package version 0.1.0},
     url = {https://github.com/poissonconsulting/checkr},
   }
 ```
