+++ revdep/timing.md
@@ -1,0 +1,7 @@
+# Check times
+
+|package |version | check_time|
+|:-------|:-------|----------:|
+|rpdo    |0.2.3   |       16.7|
+
+
