+++ revdep/problems.md
@@ -1,0 +1,27 @@
+# Setup
+
+## Platform
+
+|setting  |value                                       |
+|:--------|:-------------------------------------------|
+|version  |R version 3.4.3 Patched (2017-12-05 r73849) |
+|system   |x86_64, darwin15.6.0                        |
+|ui       |RStudio (1.1.423)                           |
+|language |(EN)                                        |
+|collate  |en_CA.UTF-8                                 |
+|tz       |America/Vancouver                           |
+|date     |2018-03-02                                  |
+
+## Packages
+
+|package |*  |version |date       |source                              |
+|:-------|:--|:-------|:----------|:-----------------------------------|
+|checkr  |*  |0.1.0   |2018-03-02 |local (poissonconsulting/checkr@NA) |
+
+# Check results
+
+0 packages with problems
+
+
+
+
