### 2018-03-09 #563 @arfon
@joethorley - many thanks for the response. I agree this is similar to some other packages/submissions.

Please go ahead and modify the paper to better fit with the submission guidelines: http://joss.theoj.org/about#paper_structure
