### 2018-03-13 2329cae
added check_flag_na() to check is scalar logical
