+++ tests/testthat/test-flag.R
@@ -11,3 +11,16 @@ test_that("check_flag errors", {
   expect_error(check_flag(y), "y must not include missing values")
   expect_identical(check_flag(FALSE), FALSE)
 })
+
+test_that("check_flag_na errors", {
+  y <- TRUE
+  expect_identical(check_flag_na(y), y)
+  y <- 2
+  expect_error(check_flag_na(y), "y must be class logical")
+  y <- c(TRUE, TRUE)
+  expect_error(check_flag_na(y), "y must have 1 element")
+  y <- NA
+  expect_identical(check_flag_na(y), y)
+  expect_identical(check_flag_na(FALSE), FALSE)
+})
+
