+++ NAMESPACE
@@ -8,6 +8,7 @@ export(check_date)
 export(check_datetime)
 export(check_environment)
 export(check_flag)
+export(check_flag_na)
 export(check_function)
 export(check_inherits)
 export(check_join)
