### 2018-03-13 3bce41c
update JOSS paper
