+++ inst/JOSS/paper.bib
@@ -1,4 +1,19 @@
-
+ @Manual{thorley_rpdo_2018,
+    title = {rpdo: Pacific Decadal Oscillation Index Data},
+    author = {Joe Thorley and Nathan Mantua and Steven R. Hare},
+    year = {2018},
+    note = {R package version 0.2.3},
+    url = {https://cran.r-project.org/web/packages/checkr/index.html},
+  }
+  
+@Manual{thorley_mcmcr_2018,
+    title = {mcmcr: Manipulate MCMC Samples},
+    author = {Joe Thorley},
+    year = {2018},
+    note = {R package version 0.0.0.9049},
+  	url = {https://github.com/poissonconsulting/mcmcr}
+  }
+  
 @incollection{hutchison_evaluating_2012,
 	address = {Berlin, Heidelberg},
 	title = {Evaluating the {Design} of the {R} {Language}},
