### 2018-03-13 #563 @joethorley
I've modified the paper to better fit with the submission guidelines. Possible reviewers include @wrathematics and @boshek.
