### 2018-03-13 0139a0a
added `check_homogenous()` to check object's elements are the same class
