+++ NAMESPACE
@@ -10,6 +10,7 @@ export(check_environment)
 export(check_flag)
 export(check_flag_na)
 export(check_function)
+export(check_homogenous)
 export(check_inherits)
 export(check_join)
 export(check_key)
