### 2018-03-13 #563 @boshek
@arfon Just trying to gauge whether I am suitable reviewer. Is [this](https://github.com/poissonconsulting/checkr/blob/master/inst/JOSS/paper.md) the extent of the paper? Or am I missing something?
