### 2018-03-13 #563 @arfon
> @arfon Just trying to gauge whether I am suitable reviewer. Is this the extent of the paper? Or am I missing something?

Yes that's it. (It could probably do with a little more detail 😁).
