### 2018-03-14 976dbac
added `check_unused()` to check `...` is unused within a function
