+++ tests/testthat/test-unused.R
@@ -1,0 +1,7 @@
+context("unused")
+
+test_that("check_unused errors", {
+  fun <- function(...) check_unused(...)
+  expect_null(fun())
+  expect_error(fun(1), "... must be unused")
+})
