+++ NAMESPACE
@@ -37,5 +37,6 @@ export(check_string)
 export(check_tz)
 export(check_unique)
 export(check_unnamed)
+export(check_unused)
 export(check_vector)
 export(checkor)
