+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: checkr
 Title: Check the Properties of Common R Objects
-Version: 0.1.0.9001
+Version: 0.1.0.9002
 Date: 2018-03-13
 Authors@R: person("Joe", "Thorley", email = "joe@poissonconsulting.ca", role = c("aut", "cre"))
 Description: Expressive assertive pipe-friendly functions to check the properties of common R objects.
