### 2018-03-14 #563 @boshek
@arfon Happy to then. Do you have any reviewing specs I can peruse?

@joethorley I think we likely can draw from [here](https://github.com/poissonconsulting/checkr/blob/master/vignettes/assertive-programming.Rmd) as well.
