### 2018-03-14 #624 @arfon
@joethorley - this PR will fix the paper compilation issues: https://github.com/poissonconsulting/checkr/pull/6
