### 2018-03-22 #624 @arfon
Thanks for the thorough and thoughtful review @boshek. @joethorley - please let me know when you've had a chance to incorporate @boshek's review feedback.
