### 2018-03-22 b5013eb
add x_name = "..." to check_unused
