+++ R/unused.R
@@ -3,6 +3,7 @@
 #' Checks whether ... is unused. It can only be used in functions.
 #' 
 #' @param ... The arguments to check.
+#' @param x_name A string of the name of the object.
 #' @param error A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.
 #' @return An invisible copy of x (if it doesn't throw an error).
 #' @export
@@ -10,8 +11,9 @@
 #' fun <- function(...) check_unused(..., error = FALSE)
 #' fun()
 #' fun(1)
-check_unused <- function(..., error = TRUE) {
+check_unused <- function(..., x_name = "...", error = TRUE) {
+  check_string_internal(x_name)
   if(length(list(...)))
-    on_fail("... must be unused", error = error)
+    on_fail(x_name, " must be unused", error = error)
   invisible(...)
 }
