### 2018-03-22 1e53fe3
replace 1:nrow(x) with seq_len(nrow(x))
