+++ R/data.R
@@ -3,7 +3,7 @@
 #' @inheritParams check_nrow
 #' @param x The object to check.
 #' @param values An optional character vector specifying the column names or a named list specifying the column names and values.
-#' @param exclusive A flag indicating whether other columns are permitted.
+#' @param exclusive A flag indicating whether other columns are not permitted.
 #' @param order A flag indicating whether the columns have to occur in the same order as values.
 #' @param key A character vector of the columns that represent a unique key.
 #' @param x_name A string of the name of the object.
