### 2018-03-22 5bf2293
added package level documentation
