+++ man/checkr-package.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/checkr-package.R
+\docType{package}
+\name{checkr-package}
+\alias{checkr}
+\alias{checkr-package}
+\title{checkr: Check the Properties of Common R Objects}
+\description{
+Expressive assertive pipe-friendly functions to check the properties of common R objects.
+In the case of failure the functions issue informative error messages that
+(hopefully!) allow users to quickly identify and fix any problems.
+The functions are designed to be used in scripts and packages.
+}
+\seealso{
+Useful links:
+\itemize{
+  \item \url{https://github.com/poissonconsulting/checkr}
+  \item Report bugs at \url{https://github.com/poissonconsulting/checkr/issues}
+}
+
+}
+\author{
+\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca}
+
+}
+\keyword{internal}
