### 2018-03-22 3880b0f
add recommendation for checkr versus alternatives
