+++ vignettes/assertive-programming.Rmd
@@ -37,6 +37,8 @@ The most useful packages that I am aware of that satisfy these criteria are
 
 This vignette consists of an overview of each package followed by a comparison.
 
+In short the `checkr` package is recommended over the alternatives if you are looking for a set of expressive, dependency-free, pipe-friendly assertive functions with customisable object names.
+
 ## Package Overview
 
 ### base
@@ -109,7 +111,7 @@ detach("package:checkmate")
 
 ### checkr
 
-The [`checkr`](https://github.com/poissonconsulting/checkr) package provides a handful of expressive functions to test the values of objects.
+The [`checkr`](https://github.com/poissonconsulting/checkr) package provides a set of expressive functions to test the values of objects.
 
 ```{r, error = TRUE}
 library(checkr)
