+++ man/checkr-package.Rd
@@ -6,7 +6,8 @@
 \alias{checkr-package}
 \title{checkr: Check the Properties of Common R Objects}
 \description{
-Expressive assertive pipe-friendly functions to check the properties of common R objects.
+Expressive assertive pipe-friendly, dependency-free functions
+to check the properties of common R objects.
 In the case of failure the functions issue informative error messages that
 (hopefully!) allow users to quickly identify and fix any problems.
 The functions are designed to be used in scripts and packages.
