+++ DESCRIPTION
@@ -3,7 +3,8 @@ Title: Check the Properties of Common R Objects
 Version: 0.1.0.9003
 Date: 2018-03-22
 Authors@R: person("Joe", "Thorley", email = "joe@poissonconsulting.ca", role = c("aut", "cre"))
-Description: Expressive assertive pipe-friendly functions to check the properties of common R objects.
+Description: Expressive assertive pipe-friendly, dependency-free functions 
+  to check the properties of common R objects.
   In the case of failure the functions issue informative error messages that 
   (hopefully!) allow users to quickly identify and fix any problems.
   The functions are designed to be used in scripts and packages.
