+++ R/data.R
@@ -1,4 +1,8 @@
 #' Check Data
+#' 
+#' Checks whether an object is a data frame. Can also check
+#' the number of rows, the names and order and values of the columns as 
+#' well as whether particular columns form a unique key.
 #'
 #' @inheritParams check_nrow
 #' @param x The object to check.
@@ -9,6 +13,7 @@
 #' @param x_name A string of the name of the object.
 #' @param error A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.
 #' @return An invisible copy of x (if it doesn't throw an error).
+#' @seealso \code{\link{check_colnames}}, \code{\link{check_nrow}} and \code{\link{check_key}} 
 #' @export
 #' 
 #' @examples
