+++ man/check_colnames.Rd
@@ -14,7 +14,7 @@ check_colnames(x, colnames = character(0), exclusive = FALSE,
 
 \item{exclusive}{A flag indicating whether other columns are not permitted.}
 
-\item{order}{A flag indicating whether the columns have to occur in the same order as names.}
+\item{order}{A flag indicating whether the columns have to occur in the same order as colnames.}
 
 \item{x_name}{A string of the name of the object.}
 
@@ -24,7 +24,10 @@ check_colnames(x, colnames = character(0), exclusive = FALSE,
 An invisible copy of x (if it doesn't throw an error).
 }
 \description{
-Check Colnames
+Checks the colnames of a data frame.
+}
+\details{
+The column names of an object can be returned using the `colnames()` function.
 }
 \examples{
 data <- data.frame(x = 1, y = 2, z = 0)
@@ -32,4 +35,7 @@ check_colnames(data, c("y", "x"), error = FALSE)
 check_colnames(data, c("y", "x"), exclusive = TRUE, error = FALSE)
 check_colnames(data, c("y", "x"), order = TRUE, error = FALSE)
 check_colnames(data, c("a"), error = FALSE)
+}
+\seealso{
+\code{\link{check_data}}
 }
