+++ README.md
@@ -22,7 +22,8 @@ assertive functions to check the properties of common R objects.
 In the case of failure the functions, which are designed to be used in
 scripts and packages, issue informative error messages.
 
-For a comparison with similar packages see the `assertive-programming`
+For an overview of the functions see the `checkr-naming` vignette and
+for a comparison with similar packages see the `assertive-programming`
 vignette.
 
 ## Demonstration
