### 2018-03-26 2aa57f9
add documentation and deprecated check_tz for check_tzone
