+++ R/key.R
@@ -1,7 +1,6 @@
 #' Check Key
 #'
 #' Checks that columns in a data frame represent a unique key.
-#'
 #' By default all the columns are checked.
 #'
 #' @param x The data to check.
@@ -9,6 +8,7 @@
 #' @param x_name A string of the name of the object.
 #' @param error A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.
 #' @return An invisible copy of x (if it doesn't throw an error).
+#' @seealso \code{\link{check_data}}
 #' @export
 #'
 #' @examples
