+++ R/pattern.R
@@ -8,6 +8,7 @@
 #' @param x_name A string of the name of the object.
 #' @param error A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.
 #' @return An invisible copy of x (if it doesn't throw an error).
+#' @seealso \code{\link{check_nchar}} and \code{\link{check_regex}}
 #' @export
 #'
 #' @examples
