+++ R/levels.R
@@ -1,4 +1,7 @@
 #' Check Levels
+#' 
+#' Checks the levels in a factor including the order and 
+#' whether other levels are permitted.
 #'
 #' @param x The object to check.
 #' @param levels A character vector of the levels.
@@ -8,6 +11,7 @@
 #' @param x_name A string of the name of the object.
 #' @param error A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.
 #' @return An invisible copy of x (if it doesn't throw an error).
+#' @seealso \code{\link{check_nlevels}} and \code{\link{check_vector}}
 #' @export
 #'
 #' @examples
