+++ R/unnamed.R
@@ -1,9 +1,12 @@
 #' Check Unnamed
+#' 
+#' Checks whether an objects is unnamed.
 #'
 #' @param x The object to check.
 #' @param x_name A string of the name of the object.
 #' @param error A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.
 #' @return An invisible copy of x (if it doesn't throw an error).
+#' @seealso \code{\link{check_named}}, \code{\link{check_names}} and \code{\link{check_missing_names}}
 #' @export
 #'
 #' @examples
