+++ R/colnames.R
@@ -1,8 +1,7 @@
 #' Check Colnames
 #' 
-#' Checks the colnames of a data frame.
-#' 
-#' The column names of an object can be returned using the `colnames()` function.
+#' Checks the column names of a data frame as returned by the `colnames()` function. 
+#' The function can check the order of the columns and whether other columns are permitted.
 #'
 #' @param x The data to check.
 #' @param colnames A character vector of the column names.
@@ -11,7 +10,7 @@
 #' @param x_name A string of the name of the object.
 #' @param error A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.
 #' @return An invisible copy of x (if it doesn't throw an error).
-#' @seealso \code{\link{check_data}}
+#' @seealso \code{\link{check_missing_colnames}} and \code{\link{check_data}}
 #' @export
 #'
 #' @examples
