+++ R/named.R
@@ -1,4 +1,6 @@
 #' Check Named
+#' 
+#' Checks whether an object is named.
 #'
 #' @param x The object to check.
 #' @param nchar A count or count range of the number of characters.
@@ -7,6 +9,7 @@
 #' @param x_name A string of the name of the object.
 #' @param error A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.
 #' @return An invisible copy of x (if it doesn't throw an error).
+#' @seealso \code{\link{check_unnamed}}, \code{\link{check_names}} and \code{\link{check_missing_names}}
 #' @export
 #'
 #' @examples
