### 2018-03-26 af8f2c8
add recommendations and example in JOSS
