### 2018-03-26 92df84e
simpler JOSS example
