### 2018-03-26 #624 @joethorley
@arfon I've incorporated @boshek 's review feedback. In particular I've

- revised the `exclusive` argument documentation so that it reads "A flag indicating whether other names are not permitted."

- added a description for each function

- generated package level documentation using `usethis::use_package_doc()` - this is a helpful function I was not aware of.

- replaced `1:nrow(x)` with `seq_len(nrow(x))`

- add a vignette `checkr-naming` which provides an overview of the functions and their naming scheme and should serve as a starting point.

- both the DESCRIPTION and JOSS paper state that the functions are designed to be used in scripts and packages.

- added a short example to the JOSS paper

- added the following text to the [vignette](https://cran.r-project.org/web/packages/checkr/vignettes/assertive-programming.html)
> In short the `checkr` package is recommended over the alternatives if you are looking for a set of expressive, dependency-free, pipe-friendly assertive functions with customisable object names.

- added a .github folder with customized ISSUE and PULL REQUEST templates - I didn't know about this - its very cool
