### 2018-03-26 #624 @boshek
So good! In particular I really like the new vignette. Everything looks great to me. I say ship it.
