### 2018-03-26 d749a33
fix .Rmd y example
