+++ README.Rmd
@@ -56,6 +56,7 @@ the values of other objects using an elegant and expressive syntax.
 
 To check the class simply pass an object of the desired class.
 ```{r, error = TRUE}
+y <- c(2,1,0,1,NA)
 check_vector(y, values = numeric(0))
 check_vector(y, values = integer(0))
 ```
