+++ README.md
@@ -1,35 +1,40 @@
----
-output: github_document
----
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-
-
- [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
- [![Travis-CI Build Status](https://travis-ci.org/poissonconsulting/checkr.svg?branch=master)](https://travis-ci.org/poissonconsulting/checkr)
-[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/poissonconsulting/checkr?branch=master&svg=true)](https://ci.appveyor.com/project/poissonconsulting/checkr)
-[![Coverage Status](https://img.shields.io/codecov/c/github/poissonconsulting/checkr/master.svg)](https://codecov.io/github/poissonconsulting/checkr?branch=master)
-[![License: MIT](https://img.shields.io/badge/License-MIT-green.svg)](https://opensource.org/licenses/MIT)
-[![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/checkr)](https://cran.r-project.org/package=checkr)
-[![CRAN Downloads](http://cranlogs.r-pkg.org/badges/grand-total/checkr)]( https://CRAN.R-project.org/package=checkr)
+[![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
+[![Travis-CI Build
+Status](https://travis-ci.org/poissonconsulting/checkr.svg?branch=master)](https://travis-ci.org/poissonconsulting/checkr)
+[![AppVeyor Build
+Status](https://ci.appveyor.com/api/projects/status/github/poissonconsulting/checkr?branch=master&svg=true)](https://ci.appveyor.com/project/poissonconsulting/checkr)
+[![Coverage
+Status](https://img.shields.io/codecov/c/github/poissonconsulting/checkr/master.svg)](https://codecov.io/github/poissonconsulting/checkr?branch=master)
+[![License:
+MIT](https://img.shields.io/badge/License-MIT-green.svg)](https://opensource.org/licenses/MIT)
+[![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/checkr)](https://cran.r-project.org/package=checkr)
+[![CRAN
+Downloads](http://cranlogs.r-pkg.org/badges/grand-total/checkr)](https://CRAN.R-project.org/package=checkr)
 
 # checkr
 
-`checkr` is an R package of assertive functions to check the properties of common R objects.
+`checkr` is an R package of assertive functions to check the properties
+of common R objects.
 
-In the case of failure the functions, which are designed to be used in scripts and packages,
-issue informative error messages.
+In the case of failure the functions, which are designed to be used in
+scripts and packages, issue informative error messages.
 
-For an overview of the functions see the `checkr-naming` vignette and for a comparison with similar packages see the `assertive-programming` vignette.
+For an overview of the functions see the `checkr-naming` vignette and
+for a comparison with similar packages see the `assertive-programming`
+vignette.
 
-The `checkr` package is recommended over the alternatives if you are looking for a set of expressive, dependency-free, pipe-friendly assertive functions with customisable object names.
+The `checkr` package is recommended over the alternatives if you are
+looking for a set of expressive, dependency-free, pipe-friendly
+assertive functions with customisable object names.
 
 ## Demonstration
 
 The following code demonstrates the `check_data()` function
 
-```r
+``` r
 library(checkr)
 
 # the starwars data frame in the dplyr package fails many of these checks
@@ -53,14 +58,16 @@ check_data(dplyr::starwars, values = list(
 
 ## Values
 
[…]
