+++ NEWS.md
@@ -1,3 +1,5 @@
+# checkr 0.2.0
+
 - deprecated `check_tz()` for `check_tzone()`
 - added `check_unused()` to check `...` is unused within a function
 - added `check_homogenous()` to check object's elements are the same class
