+++ README.md
@@ -171,7 +171,7 @@ To install the latest development version from the Poisson drat
 To cite package 'checkr' in publications use:
 
   Joe Thorley (2018). checkr: Check the Properties of Common R
-  Objects. R package version 0.1.0.9004.
+  Objects. R package version 0.2.0.
   https://github.com/poissonconsulting/checkr
 
 A BibTeX entry for LaTeX users is
@@ -180,7 +180,7 @@ A BibTeX entry for LaTeX users is
     title = {checkr: Check the Properties of Common R Objects},
     author = {Joe Thorley},
     year = {2018},
-    note = {R package version 0.1.0.9004},
+    note = {R package version 0.2.0},
     url = {https://github.com/poissonconsulting/checkr},
   }
 ```
