+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: checkr
 Title: Check the Properties of Common R Objects
-Version: 0.1.0.9004
+Version: 0.2.0
 Date: 2018-03-26
 Authors@R: person("Joe", "Thorley", email = "joe@poissonconsulting.ca", role = c("aut", "cre"))
 Description: Expressive assertive pipe-friendly, dependency-free functions 
