### 2018-03-26 b4e1804
update CRAN comments
