### 2018-03-26 a47d1e8
bump version and add zenodo badge
