### 2018-03-27 #624 @boshek
@joethorley Great for me as well. Looking forward to incorporating checkr in some future work.
